+++ client/src/Components/Community/SubjectDropdown.tsx
@@ -7,7 +7,7 @@ import { useParams } from 'react-router-dom';
 /* eslint-disable jsx-a11y/click-events-have-key-events */
 import styled from 'styled-components';
 
-import useAxios from '../../apis/customAxios';
+import useAxios from '../../hooks/useAxios';
 
 const Button = styled.button`
 	border: 1px solid rgb(214, 217, 219);
