+++ client/src/Components/UserHeader.tsx
@@ -5,7 +5,7 @@ import { HTMLAttributes, useEffect, useState } from 'react';
 import { useSelector } from 'react-redux';
 import styled from 'styled-components';
 
-import useAxios from '../apis/customAxios';
+import useAxios from '../hooks/useAxios';
 import { Iuser } from '../reducers/userInfoReducer';
 import { RootState } from '../store/Store';
