+++ client/src/pages/Community/PostUpdate.tsx
@@ -14,9 +14,9 @@ import colorSyntax from '@toast-ui/editor-plugin-color-syntax';
 import { Editor } from '@toast-ui/react-editor';
 
 import { Api } from '../../apis/customAPI';
-import useAxios from '../../apis/customAxios';
 import SearchPlace from '../../Components/Community/SearchPlace';
 import SubjectDropdown from '../../Components/Community/SubjectDropdown';
+import useAxios from '../../hooks/useAxios';
 
 const Container = styled.div`
 	width: 100vw;
